[fix] Changed param type from int to bigint for data consistency

+++ Src/Application/Interfaces/IServiceService.cs
@@ -1,3 +1,5 @@
+using System.Numerics;
+
 namespace Application.Interfaces;
 
 using Application.DTOs;
@@ -5,7 +7,7 @@ using Application.DTOs;
 public interface IServiceService
 {
     Task<IEnumerable<ServiceResponseDto>> GetAllAsync();
-    Task<ServiceResponseDto> GetByIdAsync(int id);
+    Task<ServiceResponseDto> GetByIdAsync(BigInteger id);
     Task<ServiceResponseDto> AddAsync(ServiceRequestDto serviceRequestDto);
     Task<ServiceResponseDto> UpdateAsync(int id, ServiceRequestDto serviceRequestDto);
     Task<bool> RemoveAsync(int id);
